Extension outputs to console
This is officially version 0.1! Hooray!

I made background.js, the script that actually runs in the extension. So far, it just writes into console, but I'm happy with that, since it means my code is running at all.

I also reformatted the manifest to better fit the format used in Google's documentation.