Prettied up the popup
This update adds a bearable color scheme to the popup, some aesthetic changes, and some animation for the button. I was hoping to commit some JS changes today as well, but there are too many glitches for it to be likely for me to upload the thing today. Hopefully I'll get out version 1.0 of this extension tomorrow, after the school opening ceremony.