Made icon switch to "active" or "inactive" variations based on website.
If the website you're currently on is blocked, the icon in the extension menu/toolbar is green, and if it's inactive, the icon is red.

Some other housekeeping stuff also happened here, like formatting my code in popup.js to be correct syntax (Switching from 'var' to 'let' and 'const'). I removed some obsolete debugging code. Lastly, I switched the README from goals to what the extension is, and I deleted TODO, since the extension is done for the foreseeable future, except maybe some optimizations here and there, but that's not necessarily *planned*.